pyproject.toml
[project]
name = "auto-quartzy"
version = "0.1.0"
description = "Receipt Parser & Inventory Uploader for Quartzy"
requires-python = ">=3.10"
dependencies = [
    "gradio>=5.20.1",
    "httpx>=0.28.1",
    "langchain>=0.3.20",
    "langchain-core>=0.3.44",
    "langchain-openai>=0.3.8",
    "numpy>=2.2.3",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.10.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
